Add player's cards array and number of cards
At the beginning of a game, each player is given a set number of cards.
We want this information to stay with the player object since players
will ask each other if they have certain cards. As the game progresses,
the main player will have a better understanding of what cards each
player has. The purpose of the numberOfCards integer is to use that to
determine if the main player has "discovered" what each of a particular
player's cards are (e.g. Player A has 6 cards, and the main player knows
what 5 of those cards are, therefore the main player does not know what
1 of Player A's cards are).