Implement Add Player button
When a user presses the Add Player a dialog is shown that lets the app
user input in a player name. If that player name is not currently
playing, then it adds that player to the game. If the player is already
in the game, then a Toast message is shown to notify the user.