Enter player cards on Start Game click
When the user clicks the Start Game button, they are displayed a dialog
box that lets them choose who the main player is, what cards the main
player has, and how many cards the other players have. Once this is
complete, the Add Player and Start Game buttons are removed, and the End
Game and New Turn buttons are shown.