Switch from spinner to swipe tabbed layout
Previously, switching from the predctions tab to a player tab required
the user to click the spinner then select the tab that they want to
view. This was annoying to use since it requires multiple clicks. Using
a ViewPager layout seemed like a much better user expierence because it
lest the user simply swipe from one tab to another quickly.