Add New Turn dialogs
When the New Turn button is pressed, the user first selects which player
is asking the question, then the user selects the player being
questioned from a list of remaining players. Then the Suspect, Weapon,
and Room cards are selected for the question, and finally the last
dialog box lets the user show what was answered.