Add Player knowledge to handle getting question answer
When a Player gets an answer to one of their questions, their knowledge
gets updated. A Player keeps information on each Card including what
Player gave them knowledge on that card. If a Player gains new knowledge
from a Player on a card that already contains knowledge from that
Player, the greater of the two knowledges is kept.