Add state machine to Game
Previously, buttons in fragment_main would be reset on the occasion
when the MainActivity goes through saving and restoring state (e.g.
rotating the device). This would cause issues if the game should be in
a state where the user should be able to ask a question, or start the
game.
This adds state to the Game that gets modified when new players get
added to the game, or chooses to start the game. When the Game's state
is changed, a broadcast will be sent out to notify that the game state
has changed and to update the UI accordingly. This makes it possible to
set the UI to the proper state when the Activity restores its state.