Fix overwrite issue with Player cardPlayerKnowledge
This fixes a bug where all knowledge for a particular card would be
rewritten if knowledge on that card was added from a player that was not
yet in the HashMap for that card. This was due to an OR which would put
a new hash map with the one new entry even if there was a preexisting
HashMap for that card. This fixes that by getting the preexisting
HashMap to add the entry to if it exists. If it doesn't, then it should
create a new one.