Generate and display predictions
Whenever a question between two players is asked and answered, the total
knowledge of the game cards in play is refined. This combines all of the
players knowledge, and creates predictions for the likelihood of each
combination being the true answer. These predictions are shown on the
main tab whenever the Game sends out a broadcast to notify the UI that
there are new predictions.