Improve performance and readability of Predictions class
To help improve the Predictions class, the compareTo method has been
simplified to be easier to read and to take up less loc. All HashMaps
have been converted to SparseArrays since SparseArrays are more memory
efficient. This will help in the future when the Predictions class is
computing a prediction by using the current knowledge and stepping
through all of the game's plays.