Add a test to see if a questioned card is guilty by player
Whenever a questioned player states that they do not contain any of the
cards in question, the asking player tests to see if it is possible for
one of those cards to be guilty. This is possible because each player's
knowledge is initialized at the beginning of the game so that each
player knows of the existence of every player playing that game. This
knowledge initialization is done at the beginning of the game after all
of th players have been established.