Show player cards on player fragment
This adds the ability for the user to see what cards each player has. As
the main player discovers more information on what cards each of the
other players have, the player tabs get updated to display what cards
the main player knows each player to have. This is useful to determine
if you can get any more information out of a player.
gets updated to