Use dimens for defining layout margins
Previously, many of the values for the different margins were defined
inside of the layout xml file. This was clunky and made it difficult to
keep margins consistent. Since some dimensions are already defined in
the dimens file, those dimensions are now being used to define the
margins for different UI elements. This will become the new norm for the
project and will help keep things clean.