Show player's current knowledge in player fragments
This commit adds a second list to the player fragments that shows the
respective player's knowledge. The information displayed can be used by
the user to determine the best question to ask. In the future, the app
will use the different player's knowledge to determine what question
is the optimal question to gain the most useful information from
specific players.