Implement "End Game" button functionality
When the user presses the "End Game" button they are presented with a
confirmation dialog. If the user confirms to end the game, then the
current game is reset, and the player tabs are removed.
This commit also changes how a new game is handled as it only will not
remove the main (predictions) tab and add a new main (predictions) tab,
since this is somewhat unnecessary. Now, it will only add a new main
(predictions) tab if one does not already exist.