Fix ListViews to show entire list
For some reason, if a ListView is in a ConstraintLayout that contains
other views, the ListView seems to be cut off or end past the end of the
screen. This commit places each ListView (Predictions, Player Cards, and
Player Knowledge) in their own ConstraintLayouts. This seems to fix that
problem.