Show and hide Player's cads and knowledge list on title click
The player tab show information on the player's cards and their
knowledge. These two lists can be very large causing the user to have to
do a lot of scrolling. To reduce this, the user is now able to click the
title above the ListView to either hide or show the related list. This
cleans up the Player tab and makes the app more visually pleasing.