Add new Card.Knowledge class to hold card knowledge
Previously we had been simply using different int values for card
knowledge and then have to place those in a list in the correct order to
relate knowledge to a specific card. This commit adds a Knowledge class
to Card to make it easier to relate a card to a knowledge level.