Use monospaced text style for ListView items
This commit makes it so the items in each of the three ListViews
(Predictions, Player Cards, and Player Knowledge) use a monospaced font.
Doing this makes it easier to line up predictions and knowledge so that
the user can more easily read the information. This is only temporary,
and will be changed when custom item layouts are implemented.