Fix issue with new Knowledge
Because Card.Knowledge does not have a default value, arrays that had
been created but not initialized were full of null objects. This caused
NullPointerExceptions to be thrown in the case where we were attempting
to get the knowledge level from a card in one of those lists.
This commit also changes the string printed for predictions, making it a
bit easier for the user to read.