Change Predictions class name to be more clear/concise
Previously, we had a Prediction inner class in the Predictions class.
Predictions held the static methods for creating predictions and the
Prediction class actually held the information for a prediction. This
made things a bit unclear. This commit simply cleans that up by having
a class named Prediction that actually holds the info for predictions.
It has an inner class called Utils for holding the static methods for
generating predictions.
This commit also makes it so the MainFragment actually makes use of the
predictions list passed to it during creation.