Use getView() rather than save the view that is inflated
For the Player and Main fragments a reference to the inflated view was
saved so that other views could be accessed. This is unnecessary because
fragments have a method to get the top view. Since this framework is in
place already, we should probably use that instead.
Some other minor changes were made to make better use of getView().